+++ zoom_platform_interface/lib/method_channel_zoom.dart
@@ -13,9 +13,15 @@ class MethodChannelZoom extends ZoomPlatform {
     assert(options != null);
 
     var optionMap = new Map<String, String>();
-    // optionMap.putIfAbsent("appKey", () => options.appKey);
-    // optionMap.putIfAbsent("appSecret", () => options.appSecret);
-    optionMap.putIfAbsent("jwtToken", () => options.jwtToken);
+    if (options.appKey != null) {
+      optionMap.putIfAbsent("appKey", () => options.appKey!);
+    }
+    if (options.appSecret != null) {
+      optionMap.putIfAbsent("appSecret", () => options.appSecret!);
+    }
+    if (options.jwtToken != null) {
+      optionMap.putIfAbsent("jwtToken", () => options.jwtToken!);
+    }
     optionMap.putIfAbsent("domain", () => options.domain);
     return channel
         .invokeMethod<List>('init', optionMap)