+++ zoom_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3+2
+* jwtToken to optional
+
+## 1.0.3+1
+* bug fixed
+
+## 1.0.3
+* add back appKey&appSecret
+
 ## 1.0.2
 * Add new options to use on ZoomMtg.init() (Web implementation)